Split decoration types into one module per type
Whisker's own conventions ask for one public type per module, and
`decorations.rs` had grown to three. This commit moves each into its own
file behind the same public path, with no change to any signature or any
behavior, so the change that follows can add types without making the
violation worse.
